chore: move screen interface to include
move screen interface to include

Log:
